Remove unnecessary files from hls_spatial
