+++ data_subscriber/hls_spatial/hls_spatial_catalog_connection.py
@@ -2,12 +2,12 @@ from aws_requests_auth.boto_utils import BotoAWSRequestsAuth
 from elasticsearch import RequestsHttpConnection
 from hysds.celery import app
 
-from data_subscriber.hls.hls_catalog import HLSProductCatalog
+from data_subscriber.hls_spatial.hls_spatial_catalog import HLSSpatialProductCatalog
 
 DATA_SUBSCRIBER_CONN = None
 
 
-def get_hls_catalog_connection(logger):
+def get_hls_spatial_catalog_connection(logger):
     global DATA_SUBSCRIBER_CONN
 
     if DATA_SUBSCRIBER_CONN is None:
@@ -18,7 +18,7 @@ def get_hls_catalog_connection(logger):
 
         if aws_es is True:
             aws_auth = BotoAWSRequestsAuth(aws_host=es_host, aws_region=region, aws_service='es')
-            DATA_SUBSCRIBER_CONN = HLSProductCatalog(
+            DATA_SUBSCRIBER_CONN = HLSSpatialProductCatalog(
                 es_url=es_url,
                 logger=logger,
                 http_auth=aws_auth,
@@ -31,5 +31,5 @@ def get_hls_catalog_connection(logger):
                 retry_on_timeout=True,
             )
         else:
-            DATA_SUBSCRIBER_CONN = HLSProductCatalog(es_url, logger)
+            DATA_SUBSCRIBER_CONN = HLSSpatialProductCatalog(es_url, logger)
     return DATA_SUBSCRIBER_CONN
